requirements.txt
openai
braintrust
python-dotenv
